visualizerTab with conditional rendering
